Filter-final/requirements.txt
alembic==1.0.2
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
colorama==0.4.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
numpy==1.15.3
pandas==0.23.4
Pillow==5.3.0
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pytest==3.9.3
pytest-mock==1.10.0
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
six==1.11.0
SQLAlchemy==1.2.13
Werkzeug==0.14.1
WTForms==2.2.1
